refactor(types): split interfaces.ts into named sub-interface modules; move HttpError to errors.ts
Splits the single 255-line interfaces.ts into src/types/{core,config,logger}.ts
and src/errors.ts, with interfaces.ts reduced to a re-export barrel so every
existing import path and name keeps working. The former inline Options.config
object literal becomes named, composed sub-interfaces (FormattingConfig,
OutputConfig, RedactionConfig, RequestTrackingConfig, PinoConfig) via
LogixlysiaConfig extends ..., giving each concern a named home for docs and
reducing merge-conflict surface on the highest-churn module in the package.

HttpError is a runtime class, not a type, so it moves out of the types-only
module into its own errors.ts and is re-exported as a value (not `export type`)
from interfaces.ts for compatibility.

Combined into a single commit: interfaces.ts is rewritten atomically as a
barrel, so separating the sub-interface split from the HttpError move would
require an artificial intermediate state.

No behavior change. d.ts diff against origin/main shows only structural
reshuffling: every previously exported name is still exported, no type is
narrowed, and a compile probe assigning every config key to Options still
type-checks.
