+++ packages/logixlysia/src/errors.ts
@@ -1,0 +1,8 @@
+export class HttpError extends Error {
+  readonly status: number
+
+  constructor(status: number, message: string) {
+    super(message)
+    this.status = status
+  }
+}
