+++ packages/logixlysia/src/types/config.ts
@@ -1,0 +1,186 @@
+import type { LoggerOptions as PinoLoggerOptions } from 'pino'
+import type { LogLevel } from './core'
+
+export interface Transport {
+  log: (
+    level: LogLevel,
+    message: string,
+    meta?: Record<string, unknown>
+  ) => void | Promise<void>
+}
+
+export interface LogRotationConfig {
+  compress?: boolean
+  compression?: 'gzip'
+  /**
+   * Rotate at a fixed interval, e.g. '1d', '12h'.
+   */
+  interval?: string
+  /**
+   * Keep at most N files or keep files for a duration like '7d'.
+   */
+  maxFiles?: number | string
+  /**
+   * Max log file size before rotation, e.g. '10m', '5k', or a byte count.
+   */
+  maxSize?: string | number
+}
+
+export interface LogFilter {
+  /**
+   * Array of log levels to allow. If specified, only logs with these levels will be processed.
+   * If not specified, all log levels will be allowed.
+   */
+  level?: LogLevel[]
+}
+
+/**
+ * Configuration for pino-pretty transport output.
+ *
+ * - `true`: Enable pretty printing with default options
+ * - `false` or `undefined`: Disable pretty printing
+ * - Object: Enable with custom pino-pretty options (colorize, translateTime, messageKey, errorKey, etc.)
+ *
+ * @see https://github.com/pinojs/pino-pretty#options
+ */
+export type PrettyPrintConfig = boolean | Record<string, unknown>
+
+export type LogPreset = 'dev' | 'prod' | 'json'
+
+/** Context passed to {@link Options.config.onError} when a sink fails. */
+export interface SinkErrorContext {
+  error: unknown
+  sink: 'file' | 'rotation' | 'transport'
+}
+
+export interface RequestIdConfig {
+  /**
+   * Enable request ID generation.
+   * When used as a boolean on `Options.config.requestId`, `true` enables with defaults.
+   */
+  enabled?: boolean
+  /**
+   * Custom ID generator function.
+   * @default crypto.randomUUID()
+   */
+  generator?: () => string
+  /**
+   * Header name to read from the incoming request and write to the response.
+   * @default 'X-Request-Id'
+   */
+  header?: string
+}
+
+export interface FormattingConfig {
+  /** How many object nesting levels to expand in the context tree; default 1. */
+  contextDepth?: number
+  customLogFormat?: string
+  ip?: boolean
+  /** Include query parameters in the logged URL path; default false. */
+  logQueryParams?: boolean
+  /** Service name shown in `{service}` token (e.g. evlog-style `[my-app]`). */
+  service?: string
+  /** Render `data.context` as tree lines under the main log line; default true. */
+  showContextTree?: boolean
+  showStartupMessage?: boolean
+  /** Duration (ms) below this uses green; default 500. */
+  slowThreshold?: number
+  startupMessageFormat?: 'simple' | 'banner'
+  timestamp?: {
+    translateTime?: string
+  }
+  useColors?: boolean
+  /** Duration (ms) at or above this uses red + `{speed}` badge; default 1000. */
+  verySlowThreshold?: number
+}
+
+export interface OutputConfig {
+  disableFileLogging?: boolean
+  disableInternalLogger?: boolean
+  /** Directory mode for created log directories. @default 0o700 */
+  logDirMode?: number
+  /** File mode for created log files. @default 0o600 */
+  logFileMode?: number
+  logFilePath?: string
+  logRotation?: LogRotationConfig
+  /**
+   * Called when a sink (transport, file, rotation) fails. Errors thrown by
+   * the hook itself are swallowed. When absent, failures go to stderr
+   * (rate-limited for transports).
+   */
+  onError?: (context: SinkErrorContext) => void
+  transports?: Transport[]
+  useTransportsOnly?: boolean
+}
+
+export interface RedactionConfig {
+  /**
+   * Automatically redact sensitive information (PII) from logs.
+   * Masks emails, IP addresses, Luhn-valid payment card numbers, and JWTs in strings and deeply nested objects.
+   */
+  autoRedact?: boolean
+
+  /**
+   * Log the offending payload (`found`/`errors`) from validation errors.
+   * Off by default: request bodies routinely contain credentials.
+   * @default false
+   */
+  logErrorPayload?: boolean
+
+  /**
+   * Additional key/header names (case-insensitive; `-`/`_`/camelCase variants
+   * are normalized) whose values are redacted when `autoRedact` is enabled.
+   * Extends the built-in list (authorization, cookie, x-api-key, password,
+   * secret, token, session, …).
+   */
+  redactKeys?: string[]
+}
+
+export interface RequestTrackingConfig {
+  /** Skip automatic WebSocket lifecycle logs from `wrapWs`; default false. */
+  disableWebSocketLogging?: boolean
+
+  /**
+   * Enable automatic request ID generation and propagation.
+   *
+   * - `true`: Enable with defaults (`X-Request-Id` header, `crypto.randomUUID()` generator).
+   * - `false` or `undefined`: Disabled (default).
+   * - `RequestIdConfig` object: Enable with custom header name and/or generator.
+   *
+   * When enabled, the plugin will:
+   * 1. Read `X-Request-Id` (or custom header) from the incoming request — honoring IDs set by upstream proxies.
+   * 2. Generate a new UUID if no header is present.
+   * 3. Merge `{ requestId }` into the request context (appears in logs and context tree).
+   * 4. Set the header on the outgoing response for client-side tracing.
+   */
+  requestId?: boolean | RequestIdConfig
+
+  /**
+   * Enable request-scoped logger propagation via AsyncLocalStorage.
+   * When enabled, a request-scoped logger `log` is also derived on the Elysia context.
+   * @default false
+   */
+  useAsyncLocalStorage?: boolean
+}
+
+export interface PinoConfig {
+  pino?: (PinoLoggerOptions & { prettyPrint?: PrettyPrintConfig }) | undefined
+}
+
+export interface LogixlysiaConfig
+  extends FormattingConfig,
+    OutputConfig,
+    RedactionConfig,
+    RequestTrackingConfig,
+    PinoConfig {
+  logFilter?: LogFilter
+}
+
+export interface Options {
+  config?: LogixlysiaConfig
+  /**
+   * Opinionated defaults for common environments.
+   * Explicit `config` fields override preset values.
+   */
+  preset?: LogPreset
+}
